Created account class, separate from trade_executor and related factory method Account class runs from Production = True w/o errors Account class & trade_executor simulator implementations yet to be completed
