+++ src/context_factory.py
@@ -10,7 +10,7 @@ class ContextFactory():
     def __init__(self, production: bool):
         self.production = production
 
-    def create_meta_trader(self, json_settings: dict, credentials: dict) -> IContext:
+    def create_context(self, json_settings: dict, credentials: dict) -> IContext:
         if (self.production):
             return ContextMT5(json_settings, credentials)
         else:
