+++ src/interfaces.py
@@ -41,6 +41,30 @@ class ISymbols(abc.ABC):
     def get_symbol_info_tick(self) -> dict:
         pass
 
+class ITradeExecutor(abc.ABC):
+
+    @abc.abstractmethod
+    def place_order(self) -> bool:
+        pass
+
+    @abc.abstractmethod
+    def close_position(self) -> bool:
+        pass
+
+    @abc.abstractmethod
+    def calc_lot_size(self) -> float:
+        pass
+
+class IAccount(abc.ABC):
+
+    @abc.abstractmethod
+    def get_positions(self) -> dict:
+        pass
+
+    @abc.abstractmethod
+    def get_account_balance(self) -> float:
+        pass
+
