+++ src/trade_execution_mt5.py
@@ -6,24 +6,17 @@ RISK = .02
 
 class TradeExecutorMT5():
 
-    current_balance: float # in mock implementation we track an account variable. Maybe we should do that here too
-
     current_risk_per_trade: float 
     current_lot_size: float
-    
-    current_profit: float # have we defined yet?
+    account_info: object
 
-    positions: dict
-    positions_df: pd.DataFrame # necessary ? (Note: the presentation of the dataframe is easier to read but dict is easier to work with)
-
-    def __init__(self) -> None:
+    def __init__(self, account: object) -> None:
         self.current_risk_per_trade = 0.0
         self.current_lot_size = 0.0
-        self.current_balance = self.get_account_balance()
-        # self.positions_df = pd.DataFrame() # Not used
+        self.account_info = account
 
     def calc_risk_per_trade(self) -> float:
-        self.current_risk_per_trade = self.current_balance * RISK
+        self.current_risk_per_trade = self.account_info.get_account_balance() * RISK
         return self.current_risk_per_trade
     
     def calc_lot_size(self, price) -> float:
@@ -112,32 +105,11 @@ class TradeExecutorMT5():
 
         print("2. order_send done, ", result) 
         print("closed POSITION_TICKET={}, profit {}".format(position.ticket, position.profit))
-
-    def get_positions(self) -> dict:
-        self.positions = mt5.positions_get()
-        return self.positions
     
     def close_all_positions(self, bid, ask, deviation) -> bool:
-        positions = self.get_positions()
+        positions = self.account_info.get_positions()
         for position in positions:
             self.close_position(position, bid, ask, deviation)
-            
-    # Unsure if this is necessary - easier to work with positions as a dict rather than a DataFrame
-    """
-    def get_positions_df(self) -> pd.DataFrame:
-        self.get_positions()
-        self.positions_df = pd.DataFrame(list(self.positions),columns=self.positions[0]._asdict().keys())
-        return self.positions_df
-    """
-    
-    def get_account_info(self) -> dict:
-        account_info_dict = mt5.account_info()._asdict()
-        return account_info_dict
-    
-    def get_account_balance(self) -> float:
-        account_info_dict = self.get_account_info()
-        balance = account_info_dict['balance']
-        return balance
     
     def do_nothing(self) -> None:
         print("No actionable trades!")
